requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1
apscheduler==3.10.4